Move grpc + web configuration to common, simplify somewhat
