+++ jwt-server/spring/jwt-spring-common/src/main/java/org/entur/jwt/spring/decode/ClosableJwtDecoders.java
@@ -1,4 +1,4 @@
-package org.entur.jwt.spring.config;
+package org.entur.jwt.spring.decode;
 
 import org.springframework.security.oauth2.jwt.JwtDecoder;
