+++ jwt-server/spring/jwt-spring-grpc-native/src/test/java/org/entur/jwt/spring/grpc/netty/FastIssuerAndCachedJwtDecoderContextTest.java
@@ -7,6 +7,7 @@ import org.entur.jwt.junit5.AccessToken;
 import org.entur.jwt.junit5.AuthorizationServer;
 import org.entur.jwt.spring.JwkSourceMap;
 import org.entur.jwt.spring.cache.DecodedJwtCacheJwtDecoder;
+import org.entur.jwt.spring.decode.FastIssuerJwtDecoder;
 import org.entur.jwt.spring.grpc.AbstractGrpcTest;
 import org.entur.jwt.spring.grpc.test.GreetingResponse;
 import org.junit.jupiter.api.BeforeEach;
