+++ jwt-server/spring/jwt-spring-grpc-native/src/test/java/org/entur/jwt/spring/grpc/netty/FastIssuerJwtDecoderContextTest.java
@@ -2,6 +2,7 @@ package org.entur.jwt.spring.grpc.netty;
 
 import org.entur.jwt.junit5.AccessToken;
 import org.entur.jwt.junit5.AuthorizationServer;
+import org.entur.jwt.spring.decode.FastIssuerJwtDecoder;
 import org.entur.jwt.spring.decode.JwtHeaderToIssuerMapper;
 import org.entur.jwt.spring.grpc.AbstractGrpcTest;
 import org.entur.jwt.spring.grpc.test.GreetingResponse;
